// src/types.ts
export namespace Types {
  export type TemplateData = {
    Template_id: string
    Version: string
    Template: string
  }
  export type CategoriesData = {
    value: string
  }
  export type FieldsData = {
    value: string
  }
  export type UpdateTemplateData = {
    name: string,
    fields: string,
    category: string
  }
// ...
  export type LogRadioState = { both: boolean; templates: boolean; variables: boolean }
}